Drop "Demo" suffix from new examples
