+++ examples/java/src/main/java/org/apache/beam/examples/ApproximateQuantilesExample.java
@@ -30,7 +30,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 // beam-playground:
-//   name: ApproximateQuantilesDemo
+//   name: ApproximateQuantiles
 //   description: Demonstration of ApproximateQuantiles transform usage.
 //   multifile: false
 //   default_example: false
