+++ examples/java/src/main/java/org/apache/beam/examples/CoGroupByKeyExample.java
@@ -33,7 +33,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 // beam-playground:
-//   name: CoGroupByKeyDemo
+//   name: CoGroupByKey
 //   description: Demonstration of CoGroupByKey transform usage.
 //   multifile: false
 //   default_example: false
