+++ examples/java/src/main/java/org/apache/beam/examples/CountExample.java
@@ -29,7 +29,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 // beam-playground:
-//   name: CountDemo
+//   name: Count
 //   description: Demonstration of Count transform usage.
 //   multifile: false
 //   default_example: false
