+++ examples/java/src/main/java/org/apache/beam/examples/DistinctExample.java
@@ -29,7 +29,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 // beam-playground:
-//   name: DistinctDemo
+//   name: Distinct
 //   description: Demonstration of Distinct transform usage.
 //   multifile: false
 //   default_example: false
