+++ examples/java/src/main/java/org/apache/beam/examples/GroupByKeyExample.java
@@ -30,7 +30,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 // beam-playground:
-//   name: GroupByKeyDemo
+//   name: GroupByKey
 //   description: Demonstration of GroupByKey transform usage.
 //   multifile: false
 //   default_example: false
