+++ examples/java/src/main/java/org/apache/beam/examples/MapElementsExample.java
@@ -30,7 +30,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 // beam-playground:
-//   name: MapElementsDemo
+//   name: MapElements
 //   description: Demonstration of MapElements transform usage.
 //   multifile: false
 //   default_example: false
