+++ examples/java/src/main/java/org/apache/beam/examples/MeanPerKeyExample.java
@@ -30,7 +30,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 // beam-playground:
-//   name: MeanPerKeyDemo
+//   name: MeanPerKey
 //   description: Demonstration of Mean.perKey transform usage.
 //   multifile: false
 //   default_example: false
