+++ examples/java/src/main/java/org/apache/beam/examples/MinPerKeyExample.java
@@ -30,7 +30,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 // beam-playground:
-//   name: MinPerKeyDemo
+//   name: MinPerKey
 //   description: Demonstration of Min.perKey transform usage.
 //   multifile: false
 //   default_example: false
