+++ examples/java/src/main/java/org/apache/beam/examples/SumPerKeyExample.java
@@ -30,7 +30,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 // beam-playground:
-//   name: SumPerKeyDemo
+//   name: SumPerKey
 //   description: Demonstration of Sum.integersPerKey transform usage.
 //   multifile: false
 //   default_example: false
