+++ examples/java/src/main/java/org/apache/beam/examples/ValuesExample.java
@@ -30,7 +30,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 // beam-playground:
-//   name: ValuesDemo
+//   name: Values
 //   description: Demonstration of Values transform usage.
 //   multifile: false
 //   default_example: false
