+++ examples/java/src/main/java/org/apache/beam/examples/ViewExample.java
@@ -32,7 +32,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 // beam-playground:
-//   name: ViewDemo
+//   name: View
 //   description: Demonstration of View transform usage.
 //   multifile: false
 //   default_example: false
