+++ learning/beamdoc/CombinePerKeyExample.java
@@ -15,24 +15,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.examples;
-
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
+import org.apache.beam.sdk.transforms.Combine;
 import org.apache.beam.sdk.transforms.Create;
 import org.apache.beam.sdk.transforms.DoFn;
 import org.apache.beam.sdk.transforms.ParDo;
-import org.apache.beam.sdk.transforms.SerializableFunction;
-import org.apache.beam.sdk.transforms.WithKeys;
+import org.apache.beam.sdk.transforms.Sum;
 import org.apache.beam.sdk.values.KV;
 import org.apache.beam.sdk.values.PCollection;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 // beam-playground:
-//   name: WithKeysDemo
-//   description: Demonstration of WithKeys transform usage.
+//   name: CombinePerKey
+//   description: Demonstration of Combine.perKey transform usage.
 //   multifile: false
 //   default_example: false
 //   context_line: 48
@@ -41,27 +39,31 @@ import org.slf4j.LoggerFactory;
 //   complexity: BASIC
 //   tags:
 //     - transforms
-//     - strings
-//     - pairs
+//     - numbers
+//     - distinct
 
-public class WithKeysExample {
+public class CombinePerKeyExample {
   public static void main(String[] args) {
     PipelineOptions options = PipelineOptionsFactory.create();
     Pipeline pipeline = Pipeline.create(options);
+
     // [START main_section]
-    PCollection<String> words = pipeline.apply(Create.of("Hello", "World", "Beam", "is", "fun"));
-    PCollection<KV<Integer, String>> lengthAndWord =
-        words.apply(
-            WithKeys.of(
-                new SerializableFunction<String, Integer>() {
-                  @Override
-                  public Integer apply(String s) {
-                    return s.length();
-                  }
-                }));
+    // PCollection is grouped by key and the Double values associated with each key
+    // are combined into a Double.
+    PCollection<KV<String, Double>> salesRecords =
+        pipeline.apply(
+            Create.of(
+                KV.of("Apples", 2.0),
+                KV.of("Apples", 3.0),
+                KV.of("Apples", 5.0),
+                KV.of("Oranges", 4.0),
+                KV.of("Oranges", 8.0)));
+    PCollection<KV<String, Double>> totalSalesPerPerson =
+        salesRecords.apply(Combine.perKey(Sum.ofDoubles()));
     // [END main_section]
-    lengthAndWord.apply(
-        ParDo.of(new LogOutput<>("PCollection elements after WithKeys transform: ")));
+    // Log values
+    totalSalesPerPerson.apply(
+        ParDo.of(new LogOutput<>("PCollection numbers after Combine.perKey transform: ")));
     pipeline.run();
   }
