+++ learning/beamdoc/FlattenExample.java
@@ -15,57 +15,48 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.examples;
-
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
-import org.apache.beam.sdk.transforms.Combine;
 import org.apache.beam.sdk.transforms.Create;
 import org.apache.beam.sdk.transforms.DoFn;
+import org.apache.beam.sdk.transforms.Flatten;
 import org.apache.beam.sdk.transforms.ParDo;
-import org.apache.beam.sdk.transforms.Sum;
-import org.apache.beam.sdk.values.KV;
 import org.apache.beam.sdk.values.PCollection;
+import org.apache.beam.sdk.values.PCollectionList;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 // beam-playground:
-//   name: CombinePerKeyDemo
-//   description: Demonstration of Combine.perKey transform usage.
+//   name: Flatten
+//   description: Demonstration of Flatten transform usage.
 //   multifile: false
 //   default_example: false
-//   context_line: 48
+//   context_line: 46
 //   categories:
 //     - Core Transforms
 //   complexity: BASIC
 //   tags:
 //     - transforms
 //     - numbers
-//     - distinct
 
-public class CombinePerKeyExample {
+public class FlattenExample {
   public static void main(String[] args) {
     PipelineOptions options = PipelineOptionsFactory.create();
     Pipeline pipeline = Pipeline.create(options);
-
     // [START main_section]
-    // PCollection is grouped by key and the Double values associated with each key
-    // are combined into a Double.
-    PCollection<KV<String, Double>> salesRecords =
-        pipeline.apply(
-            Create.of(
-                KV.of("Apples", 2.0),
-                KV.of("Apples", 3.0),
-                KV.of("Apples", 5.0),
-                KV.of("Oranges", 4.0),
-                KV.of("Oranges", 8.0)));
-    PCollection<KV<String, Double>> totalSalesPerPerson =
-        salesRecords.apply(Combine.perKey(Sum.ofDoubles()));
+    // Flatten takes a PCollectionList of PCollection objects of a given type.
+    // Returns a single PCollection that contains all of the elements in the PCollection objects in
+    // that list.
+    PCollection<String> pc1 = pipeline.apply(Create.of("Hello"));
+    PCollection<String> pc2 = pipeline.apply(Create.of("World", "Beam"));
+    PCollection<String> pc3 = pipeline.apply(Create.of("Is", "Fun"));
+    PCollectionList<String> collections = PCollectionList.of(pc1).and(pc2).and(pc3);
+
+    PCollection<String> merged = collections.apply(Flatten.pCollections());
     // [END main_section]
     // Log values
-    totalSalesPerPerson.apply(
-        ParDo.of(new LogOutput<>("PCollection numbers after Combine.perKey transform: ")));
+    merged.apply(ParDo.of(new LogOutput<>("PCollection numbers after Flatten transform: ")));
     pipeline.run();
   }
