+++ learning/katas/java/Core Transforms/GroupByKey/GroupByKey/src/org/apache/beam/learning/katas/coretransforms/groupbykey/Task.java
@@ -19,7 +19,7 @@
 package org.apache.beam.learning.katas.coretransforms.groupbykey;
 
 // beam-playground:
-//   name: GroupByKey
+//   name: GroupByKeyKata
 //   description: Task from katas that groups words by its first letter.
 //   multifile: false
 //   context_line: 43
